requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
requests~=2.32.3
coverage==7.6.0
setuptools==72.1.0
drf-yasg==1.21.7
flake8==7.1.1
pytz~=2024.1
django-filter==24.3
django-environ==0.11.2
black==23.9.1